requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
python-dotenv==1.0.1
openai==1.12.0
pydub==0.25.1
SpeechRecognition==3.10.1
pytesseract==0.3.10
Pillow==10.2.0
numpy==1.26.4
python-multipart==0.0.9
langchain==0.1.9
langchain-openai==0.0.5
langchain-community==0.0.27
chromadb==0.4.22 